refactor: rename folders to singular instead of plural
