+++ src/domain/faucet/entity/error.ts
