+++ src/domain/faucet/usecase/ask-tokens/actionCreators.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 import type { ActionsUnion } from '../../../helpers/store.helper'
 import { createAction } from '../../../helpers/store.helper'
-import type { ValidationError, GatewayError } from 'domain/faucet/entities/error'
+import type { ValidationError, GatewayError } from 'domain/faucet/entity/error'
 import type { DeepReadonly } from '../../../../superTypes'
 
 export const AskTokensActions = {
